tests/wpt/web-platform-tests/html/semantics/embedded-content/the-img-element/original-referrer-policy-applied.sub.html

<!DOCTYPE html>
<head>
  <title>Deferred iframes and images with loading='lazy' use the original referrer-policy specified at the parse time</title>
  <link rel="author" title="Raj T" href="mailto:rajendrant@chromium.org">
  <link rel="help" href="https://github.com/scott-little/lazyload">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
</head>

<script>
  const below_viewport_iframe = new ElementLoadPromise("below_viewport_iframe");
  const below_viewport_img = new ElementLoadPromise("below_viewport_img");

  // Change the referrer-policy and scroll down to load the deferred elements.
  window.addEventListener("load", () => {
    below_viewport_iframe.element().referrerPolicy = "no-referrer";
    below_viewport_img.element().referrerPolicy = "no-referrer";
    document.getElementById("below_viewport_iframe").scrollIntoView();
  });

  async_test(function(t) {
    below_viewport_iframe.promise.then(
      t.step_func_done(function() {
        // The referer header should be the full URL (as specified in the iframe
        // at parse time), and not the origin (as specified in meta referrer
        // tag) or null (as overridden by iframe referrerpolicy=no-referrer).
        assert_true(below_viewport_iframe.element().contentDocument.body.innerHTML
            .includes("Referer: http://{{location[host]}}{{location[path]}}"));
      }));
  }, "Test that when deferred iframe is loaded, it uses the referrer-policy specified at parse time.");

  async_test(function(t) {
    below_viewport_img.promise.then(
      t.step_func_done(function() {
        // The image will load successfully if the full URL is sent as referrer.
        assert_true(below_viewport_img.element().complete);
        assert_greater_than(below_viewport_img.element().naturalWidth, 0);
      })
    ).catch(t.unreached_func("The image load should not fail, by sending the wrong referer header."));
  }, "Test that when deferred img is loaded, it uses the referrer-policy specified at parse time.");
</script>

<body>
  <meta name="referrer" content="origin">
  <div style="height:10000px;"></div>
  <iframe id="below_viewport_iframe" src="/xhr/resources/echo-headers.py" loading="lazy" width="200px" height="100px" referrerpolicy="unsafe-url" onload="below_viewport_iframe.resolve();">
  </iframe>
  <img id="below_viewport_img" src="resources/referrer-checker-img.py?expected_referrer=http://{{location[host]}}{{location[path]}}"
       loading="lazy" referrerpolicy="unsafe-url" onload="below_viewport_img.resolve();" onerror="below_viewport_img.reject();">
</body>
